+++ sshd-core/src/main/java/org/apache/sshd/server/shell/InvertedShell.java
@@ -22,10 +22,10 @@ import java.io.InputStream;
 import java.io.OutputStream;
 
 import org.apache.sshd.common.session.SessionHolder;
-import org.apache.sshd.server.SessionAware;
-import org.apache.sshd.server.channel.ChannelSession;
+import org.apache.sshd.server.channel.ServerChannelSessionHolder;
 import org.apache.sshd.server.command.CommandLifecycle;
 import org.apache.sshd.server.session.ServerSession;
+import org.apache.sshd.server.session.ServerSessionAware;
 import org.apache.sshd.server.session.ServerSessionHolder;
 
 /**
@@ -37,20 +37,14 @@ import org.apache.sshd.server.session.ServerSessionHolder;
  */
 public interface InvertedShell
         extends SessionHolder<ServerSession>,
-        ServerSessionHolder,
+        ServerSessionHolder, ServerChannelSessionHolder,
         CommandLifecycle,
-        SessionAware {
+        ServerSessionAware {
 
     @Override
     default ServerSession getSession() {
         return getServerSession();
     }
-
-    /**
-     * @return The {@link ChannelSession} instance through which the shell was created - may be {@code null} if shell
-     *         not started yet
-     */
-    ChannelSession getChannelSession();
 
     /**
      * Returns the output stream used to feed the shell. This method is called after the shell has been started.