+++ sshd-core/src/main/java/org/apache/sshd/server/shell/InvertedShellWrapper.java
@@ -34,10 +34,10 @@ import org.apache.sshd.common.util.threads.ThreadUtils;
 import org.apache.sshd.core.CoreModuleProperties;
 import org.apache.sshd.server.Environment;
 import org.apache.sshd.server.ExitCallback;
-import org.apache.sshd.server.SessionAware;
 import org.apache.sshd.server.channel.ChannelSession;
 import org.apache.sshd.server.command.Command;
 import org.apache.sshd.server.session.ServerSession;
+import org.apache.sshd.server.session.ServerSessionAware;
 
 /**
  * A shell implementation that wraps an instance of {@link InvertedShell} as a {@link Command}. This is useful when
@@ -46,7 +46,7 @@ import org.apache.sshd.server.session.ServerSession;
  *
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-public class InvertedShellWrapper extends AbstractLoggingBean implements Command, SessionAware {
+public class InvertedShellWrapper extends AbstractLoggingBean implements Command, ServerSessionAware {
 
     private final InvertedShell shell;
     private final Executor executor;
@@ -203,7 +203,7 @@ public class InvertedShellWrapper extends AbstractLoggingBean implements Command
         } catch (Throwable e) {
             boolean debugEnabled = log.isDebugEnabled();
             try {
-                shell.destroy(shell.getChannelSession());
+                shell.destroy(shell.getServerChannelSession());
             } catch (Throwable err) {
                 warn("pumpStreams({}) failed ({}) to destroy shell: {}",
                         this, e.getClass().getSimpleName(), e.getMessage(), e);