+++ sshd-sftp/src/main/java/org/apache/sshd/sftp/server/SftpSubsystem.java
@@ -64,12 +64,11 @@ import org.apache.sshd.common.util.io.IoUtils;
 import org.apache.sshd.common.util.threads.CloseableExecutorService;
 import org.apache.sshd.common.util.threads.ExecutorServiceCarrier;
 import org.apache.sshd.common.util.threads.ThreadUtils;
-import org.apache.sshd.server.ChannelSessionAware;
 import org.apache.sshd.server.Environment;
 import org.apache.sshd.server.ExitCallback;
-import org.apache.sshd.server.SessionAware;
 import org.apache.sshd.server.channel.ChannelDataReceiver;
 import org.apache.sshd.server.channel.ChannelSession;
+import org.apache.sshd.server.channel.ChannelSessionAware;
 import org.apache.sshd.server.command.AsyncCommand;
 import org.apache.sshd.server.command.AsyncCommandErrorStreamAware;
 import org.apache.sshd.server.command.Command;
@@ -87,8 +86,8 @@ import org.apache.sshd.sftp.common.SftpHelper;
  */
 public class SftpSubsystem
         extends AbstractSftpSubsystemHelper
-        implements Command, Runnable, SessionAware, FileSystemAware, ExecutorServiceCarrier,
-        AsyncCommand, ChannelSessionAware, ChannelDataReceiver {
+        implements Command, Runnable, FileSystemAware, ExecutorServiceCarrier,
+        AsyncCommand, ChannelDataReceiver {
 
     protected static final Buffer CLOSE = new ByteArrayBuffer(null, 0, 0);
 
@@ -98,7 +97,6 @@ public class SftpSubsystem
     protected final Map<String, Handle> handles = new ConcurrentHashMap<>();
     protected final Buffer buffer = new ByteArrayBuffer(1024);
     protected final BlockingQueue<Buffer> requests = new LinkedBlockingQueue<>();
-    protected final ChannelDataReceiver errorDataChannelReceiver;
 
     protected ExitCallback callback;
     protected IoOutputStream out;
@@ -113,39 +111,14 @@ public class SftpSubsystem
     protected int version;
 
     protected ServerSession serverSession;
-    protected ChannelSession channelSession;
     protected CloseableExecutorService executorService;
 
     /**
+     * @param channel      The {@link ChannelSession} through which the command was received
      * @param configurator The {@link SftpSubsystemConfigurator} to use
      */
-    public SftpSubsystem(SftpSubsystemConfigurator configurator) {
-        super(configurator);
-
-        ChannelDataReceiver receiver = configurator.getErrorChannelDataReceiver();
-        if (receiver == null) {
-            errorDataChannelReceiver = new ChannelDataReceiver() {
-                @Override
-                @SuppressWarnings("synthetic-access")
-                public void close() throws IOException {
-                    if (log.isDebugEnabled()) {
-                        log.debug("stderrData({}) closing", getSession());
-                    }
-
-                }
-
-                @Override
-                @SuppressWarnings("synthetic-access")
-                public int data(ChannelSession channel, byte[] buf, int start, int len) throws IOException {
-                    if (log.isDebugEnabled()) {
-                        log.debug("stderrData({}) received {} data bytes", channel, len);
-                    }
-                    return len;
-                }
-            };
-        } else {
-            errorDataChannelReceiver = receiver;
-        }
+    public SftpSubsystem(ChannelSession channel, SftpSubsystemConfigurator configurator) {
+        super(channel, configurator);
 
         CloseableExecutorService executorService = configurator.getExecutorService();
         if (executorService == null) {
@@ -153,6 +126,40 @@ public class SftpSubsystem
         } else {
             this.executorService = executorService;
         }
+
+        initializeSessionRelatedMember(channel);
+
+        ChannelDataReceiver errorDataChannelReceiver
+                = resolveErrorDataChannelReceiver(channel, configurator.getErrorChannelDataReceiver());
+        channel.setDataReceiver(this);
+        channel.setExtendedDataWriter(errorDataChannelReceiver);
+
+        SftpErrorStatusDataHandler errHandler = getErrorStatusDataHandler();
+        if (errHandler instanceof ChannelSessionAware) {
+            ((ChannelSessionAware) errHandler).setChannelSession(channel);
+        }
+    }
+
+    protected ChannelDataReceiver resolveErrorDataChannelReceiver(ChannelSession channelSession, ChannelDataReceiver receiver) {
+        return (receiver != null) ? receiver : new ChannelDataReceiver() {
+            @Override
+            @SuppressWarnings("synthetic-access")
+            public void close() throws IOException {
+                if (log.isDebugEnabled()) {
+                    log.debug("stderrData({}) closing", getSession());
+                }
+
+            }
+
+            @Override
+            @SuppressWarnings("synthetic-access")
+            public int data(ChannelSession channel, byte[] buf, int start, int len) throws IOException {
+                if (log.isDebugEnabled()) {
+                    log.debug("stderrData({}) received {} data bytes", channel, len);
+                }
+                return len;
+            }
+        };
     }
 
     @Override
@@ -170,16 +177,15 @@ public class SftpSubsystem
         return executorService;
     }
 
-    @Override
-    public void setSession(ServerSession session) {
-        this.serverSession = Objects.requireNonNull(session, "No session");
+    protected void initializeSessionRelatedMember(ChannelSession channel) {
+        serverSession = Objects.requireNonNull(channel.getServerSession(), "No session associated with the channel");
 
-        FactoryManager manager = session.getFactoryManager();
+        FactoryManager manager = serverSession.getFactoryManager();
         Factory<? extends Random> factory = manager.getRandomFactory();
         this.randomizer = factory.create();
 
-        this.fileHandleSize = SftpModuleProperties.FILE_HANDLE_SIZE.getRequired(session);
-        this.maxFileHandleRounds = SftpModuleProperties.MAX_FILE_HANDLE_RAND_ROUNDS.getRequired(session);
+        this.fileHandleSize = SftpModuleProperties.FILE_HANDLE_SIZE.getRequired(channel);
+        this.maxFileHandleRounds = SftpModuleProperties.MAX_FILE_HANDLE_RAND_ROUNDS.getRequired(channel);
 
         if (workBuf.length < this.fileHandleSize) {
             workBuf = new byte[this.fileHandleSize];
@@ -189,18 +195,6 @@ public class SftpSubsystem
     @Override
     public ServerSession getServerSession() {
         return serverSession;
-    }
-
-    @Override
-    public void setChannelSession(ChannelSession session) {
-        this.channelSession = session;
-        session.setDataReceiver(this);
-        session.setExtendedDataWriter(errorDataChannelReceiver);
-
-        SftpErrorStatusDataHandler errHandler = getErrorStatusDataHandler();
-        if (errHandler instanceof ChannelSessionAware) {
-            ((ChannelSessionAware) errHandler).setChannelSession(session);
-        }
     }
 
     @Override
@@ -242,8 +236,9 @@ public class SftpSubsystem
 
     @Override
     public void setIoOutputStream(IoOutputStream out) {
-        int channelId = channelSession.getId();
-        this.out = new BufferedIoOutputStream("sftp-out@" + channelId, channelId, out, channelSession);
+        ChannelSession channel = getServerChannelSession();
+        int channelId = channel.getId();
+        this.out = new BufferedIoOutputStream("sftp-out@" + channelId, channelId, out, channel);
     }
 
     @Override
@@ -290,22 +285,25 @@ public class SftpSubsystem
     @Override
     public void run() {
         int exitCode = 0;
+        long buffersCount = 0L;
         try {
+            ChannelSession channel = getServerChannelSession();
+            Window localWindow = channel.getLocalWindow();
             while (true) {
                 Buffer buffer = requests.take();
                 if (buffer == CLOSE) {
                     break;
                 }
                 int len = buffer.available();
+                buffersCount++;
                 process(buffer);
-                Window localWindow = channelSession.getLocalWindow();
                 localWindow.consumeAndCheck(len);
             }
         } catch (Throwable t) {
             if (!closed.get()) { // Ignore
                 Session session = getServerSession();
-                error("run({}) {} caught in SFTP subsystem: {}",
-                        session, t.getClass().getSimpleName(), t.getMessage(), t);
+                error("run({}) {} caught in SFTP subsystem after {} buffers: {}",
+                        session, t.getClass().getSimpleName(), buffersCount, t.getMessage(), t);
                 exitCode = -1;
             }
         } finally {